+++ tests/test_valhalla.py
@@ -23,7 +23,7 @@ import responses
 
 import tests as _test
 from routingpy import Valhalla
-from routingpy.direction import Direction
+from routingpy.direction import Direction, Directions
 from routingpy.expansion import Expansions
 from routingpy.isochrone import Isochrone, Isochrones
 from routingpy.matrix import Matrix
@@ -68,6 +68,38 @@ class ValhallaTest(_test.TestCase):
         self.assertEqual(routes.duration, 57)
         self.assertIsInstance(routes.geometry, list)
         self.assertIsInstance(routes.raw, dict)
+
+    @responses.activate
+    def test_directions_alternatives(self):
+        query = deepcopy(ENDPOINTS_QUERIES[self.name]["alternatives"])
+        expected = deepcopy(ENDPOINTS_EXPECTED[self.name]["alternatives"])
+
+        responses.add(
+            responses.POST,
+            "https://api.mapbox.com/valhalla/v1/route",
+            status=200,
+            json=ENDPOINTS_RESPONSES[self.name]["alternatives"],
+            content_type="application/json",
+        )
+
+        routes = self.client.directions(**query)
+
+        self.assertEqual(1, len(responses.calls))
+        self.assertEqual(json.loads(responses.calls[0].request.body.decode("utf-8")), expected)
+
+        # Since we requested alternatives, we should get a Directions object
+        self.assertIsInstance(routes, Directions)
+
+        # We should have 1 route + 1 alternative = 2 routes
+        self.assertEqual(len(routes), 2)
+
+        # Check we have stored both routes correctly
+        route1 = routes[0]
+        route2 = routes[1]
+        self.assertIsInstance(route1, Direction)
+        self.assertIsInstance(route2, Direction)
+        self.assertEqual(route1.distance, int(46.176 * 1000))
+        self.assertEqual(route2.distance, int(54.324 * 1000))
 
     @responses.activate
     def test_waypoint_generator(self):